[refactor] Move state mapping logic from views to app.jsx
