+++ src/views/conservatory.jsx
@@ -2,7 +2,6 @@ import $            from 'jquery';
 import React        from 'react';
 import { render }   from 'react-dom';
 import { Link }     from 'react-router';
-import { connect }  from 'react-redux';
 
 import Users        from '../components/users.jsx';
 import Search       from '../components/searchBar.jsx';
@@ -14,16 +13,9 @@ import Plant        from '../components/plant.jsx';
 import Modal        from '../components/plantModal.jsx';
 import Footer       from '../components/footer.jsx';
 
-import { 
-  _getGarden, 
-  _getPlants, 
-  _fetchPlant }     from '../redux/actions/helpers';
-import { 
-  toggleNewPlant }   from '../redux/actions/actions';
-
 $(() => $('[data-toggle="tooltip"]').tooltip());
 
-class Conservatory extends React.Component {
+export default class Conservatory extends React.Component {
   constructor(props) {
     super(props);
   }
@@ -44,9 +36,17 @@ class Conservatory extends React.Component {
       <div className="container-fluid">
         <div className="row search">
           <div className="column jumbotron jumbo-bg">
+
             { this.props.plants ?
-              <Search className="form-control form-control-lg" { ...this.props } /> : null
+
+              <Search className="form-control form-control-lg" { ...this.props } />
+            
+            :
+            
+              null
+
             }
+            
           </div>
         </div>
 
@@ -79,53 +79,4 @@ class Conservatory extends React.Component {
 
   }
 }
-
-function mapDispatchToProps(dispatch) {
-  return {
-    fetchGarden    : () => dispatch(_getGarden()),
-    fetchPlants   : () => dispatch(_getPlants()),
-    toggleNewPlant: () => dispatch(toggleNewPlant()),
-    fetchPlant    : (plant) => dispatch(_fetchPlant(plant))
-  };
-}
-
-function mapStateToProps(state) {
-  const user = state.get('user');
-  if( state.get('loggedIn') ){
-    return {
-      plants: state.getIn([ 'plants', 'plants' ]),
-      garden: state.getIn([ 'garden', 'garden' ]),
-      user_plants: user.get('user_plants'),
-      loggedIn: state.get('loggedIn'),
-      username: user.get('name'),
-      image: user.get('image'),
-      firstName: user.get('firstName'),
-      lastName: user.get('lastName'),
-      plantFacts: state.getIn(['plantFacts', 'plantFacts']),
-      id: user.get('id'),
-      newPlant: state.get('newPlant')
-    };
-  }
-
-  if (state.get('plantFacts') ) {
-    return {
-      plantFacts: state.getIn(['plantFacts', 'plantFacts']),
-      plants: state.getIn([ 'plants', 'plants' ]),
-      garden: state.getIn([ 'garden', 'garden' ]),
-      loggedIn: state.get('loggedIn'),
-      id: state.get('id'),
-      newPlant: state.get('newPlant')
-    };
-  }
-
-  return {
-    plants: state.getIn([ 'plants', 'plants' ]),
-    garden: state.getIn([ 'garden', 'garden' ]),
-    loggedIn: state.get('loggedIn'),
-    id: state.get('id'),
-    newPlant: state.get('newPlant')
-  };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(Conservatory);
